# api-spec.yaml
openapi: 3.0.3
info:
  title: FinanceOS API
  description: Production-grade personal finance backend API
  version: 1.0.0
  contact:
    name: FinanceOS
servers:
  - url: http://localhost:6969
    description: Local development server
...
paths:
...
  /api/v1/loans:
    get:
      tags: [Loans]
      summary: List loans
      operationId: listLoans
      security:
        - sessionAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/LoanStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Paged list of loans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageLoanResponse'
    post:
      tags: [Loans]
      summary: Create loan
      operationId: createLoan
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoanRequest'
      responses:
        '201':
          description: Loan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'
  /api/v1/loans/summary:
    get:
      tags: [Loans]
      summary: Get loans summary
      operationId: getLoansSummary
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Loans summary statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoansSummaryResponse'
  /api/v1/loans/{id}:
    get:
      tags: [Loans]
      summary: Get loan detail
      operationId: getLoanDetail
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Loan detail with events and charges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanDetailResponse'
    put:
      tags: [Loans]
      summary: Update loan
      operationId: updateLoan
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLoanRequest'
      responses:
        '200':
          description: Loan updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'
    delete:
      tags: [Loans]
      summary: Delete loan
      operationId: deleteLoan
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Loan deleted
  /api/v1/loans/{id}/close:
    post:
      tags: [Loans]
      summary: Close loan
      operationId: closeLoan
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Loan closed
  /api/v1/loans/{id}/reopen:
    post:
      tags: [Loans]
      summary: Reopen loan
      operationId: reopenLoan
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Loan reopened
  /api/v1/loans/{id}/schedule:
    get:
      tags: [Loans]
      summary: Get loan schedule
      operationId: getLoanSchedule
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Computed loan schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResultResponse'
  /api/v1/loans/{id}/events:
    post:
      tags: [Loans]
      summary: Add loan event
      operationId: addLoanEvent
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoanEventRequest'
      responses:
        '201':
          description: Event added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanEventResponse'
  /api/v1/loans/{id}/events/{eventId}:
    delete:
      tags: [Loans]
      summary: Delete loan event
      operationId: deleteLoanEvent
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Event deleted
  /api/v1/loans/{id}/payments:
    post:
      tags: [Loans]
      summary: Add loan payment
      operationId: addLoanPayment
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoanPaymentRequest'
      responses:
        '201':
          description: Payment added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanPaymentResponse'
  /api/v1/loans/{id}/payments/batch:
    post:
      tags: [Loans]
      summary: Add loan payments batch
      operationId: addLoanPaymentsBatch
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchLoanPaymentRequest'
      responses:
        '200':
          description: Batch payments added
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: integer
  /api/v1/loans/{id}/payments/{paymentId}:
    delete:
      tags: [Loans]
      summary: Delete loan payment
      operationId: deleteLoanPayment
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Payment deleted
  /api/v1/loans/{id}/charges:
    post:
      tags: [Loans]
      summary: Add loan charge
      operationId: addLoanCharge
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoanChargeRequest'
      responses:
        '201':
          description: Charge added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanChargeResponse'
  /api/v1/loans/{id}/charges/{chargeId}:
    delete:
      tags: [Loans]
      summary: Delete loan charge
      operationId: deleteLoanCharge
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: chargeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Charge deleted
  /api/v1/loans/{id}/match-suggestions:
    get:
      tags: [Loans]
      summary: Get loan match suggestions
      operationId: getLoanMatchSuggestions
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: EMI transaction match suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchSuggestionsResponse'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: FINANCEOS_SESSION
  schemas:
...
    LoanType:
      type: string
      enum: [home, car, personal, education, gold, two_wheeler, consumer_durable, other]
    RateType:
      type: string
      enum: [fixed, floating]
    LoanStatus:
      type: string
      enum: [active, closed, foreclosed]
    LoanEventType:
      type: string
      enum: [rate_change, prepayment, foreclosure]
    AdjustmentMode:
      type: string
      enum: [reduce_tenure, reduce_emi]
    LoanChargeType:
      type: string
      enum: [processing_fee, insurance_premium, foreclosure_charge, bounce_charge, late_fee, legal_valuation, other]
...
    CreateLoanRequest:
      type: object
      required: [name, lender, loanType, rateType, principal, annualRatePct, tenureMonths, startDate, firstEmiDate]
      properties:
        name: { type: string }
        lender: { type: string }
        loanType: { $ref: '#/components/schemas/LoanType' }
        rateType: { $ref: '#/components/schemas/RateType' }
        loanAccountNumber: { type: string }
        paymentAccountId: { type: string, format: uuid }
        principal: { type: number }
        annualRatePct: { type: number }
        tenureMonths: { type: integer }
        startDate: { type: string, format: date }
        firstEmiDate: { type: string, format: date }
        emiAmount: { type: number }
        notes: { type: string }
    UpdateLoanRequest:
      type: object
      properties:
        name: { type: string }
        lender: { type: string }
        loanType: { $ref: '#/components/schemas/LoanType' }
        rateType: { $ref: '#/components/schemas/RateType' }
        loanAccountNumber: { type: string }
        paymentAccountId: { type: string, format: uuid }
        principal: { type: number }
        annualRatePct: { type: number }
        tenureMonths: { type: integer }
        startDate: { type: string, format: date }
        firstEmiDate: { type: string, format: date }
        emiAmount: { type: number }
        notes: { type: string }
    LoanResponse:
      type: object
      required: [id, name, lender, loanType, rateType, principal, outstandingPrincipal, currentAnnualRatePct, currentEmi, totalInstallments, settledInstallments, startDate, firstEmiDate, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        lender: { type: string }
        loanType: { $ref: '#/components/schemas/LoanType' }
        rateType: { $ref: '#/components/schemas/RateType' }
        loanAccountNumber: { type: string }
        paymentAccountId: { type: string, format: uuid }
        principal: { type: number }
        outstandingPrincipal: { type: number }
        currentAnnualRatePct: { type: number }
        currentEmi: { type: number }
        totalInstallments: { type: integer }
        settledInstallments: { type: integer }
        startDate: { type: string, format: date }
        firstEmiDate: { type: string, format: date }
        nextDueDate: { type: string, format: date }
        projectedEndDate: { type: string, format: date }
        totalInterestPaid: { type: number }
        totalInterestRemaining: { type: number }
        effectiveAprPct: { type: number }
        status: { $ref: '#/components/schemas/LoanStatus' }
        notes: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    LoanDetailResponse:
      type: object
      required: [loan, events, charges]
      properties:
        loan: { $ref: '#/components/schemas/LoanResponse' }
        events: { type: array, items: { $ref: '#/components/schemas/LoanEventResponse' } }
        charges: { type: array, items: { $ref: '#/components/schemas/LoanChargeResponse' } }
    CreateLoanEventRequest:
      type: object
      required: [eventType, effectiveDate]
      properties:
        eventType: { $ref: '#/components/schemas/LoanEventType' }
        effectiveDate: { type: string, format: date }
        newAnnualRatePct: { type: number }
        amount: { type: number }
        adjustmentMode: { $ref: '#/components/schemas/AdjustmentMode' }
        newEmiOverride: { type: number }
        transactionId: { type: string, format: uuid }
    LoanEventResponse:
      type: object
      required: [id, loanId, eventType, effectiveDate, createdAt]
      properties:
        id: { type: string, format: uuid }
        loanId: { type: string, format: uuid }
        eventType: { $ref: '#/components/schemas/LoanEventType' }
        effectiveDate: { type: string, format: date }
        newAnnualRatePct: { type: number }
        amount: { type: number }
        adjustmentMode: { $ref: '#/components/schemas/AdjustmentMode' }
        newEmiOverride: { type: number }
        transactionId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
    CreateLoanPaymentRequest:
      type: object
      required: [installmentSeq, paymentDate, amount]
      properties:
        installmentSeq: { type: integer }
        paymentDate: { type: string, format: date }
        amount: { type: number }
        transactionId: { type: string, format: uuid }
    BatchLoanPaymentItem:
      type: object
      required: [paymentDate, amount]
      properties:
        installmentSeq: { type: integer }
        paymentDate: { type: string, format: date }
        amount: { type: number }
        transactionId: { type: string, format: uuid }
    BatchLoanPaymentRequest:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/BatchLoanPaymentItem' } }
    LoanPaymentResponse:
      type: object
      required: [id, loanId, installmentSeq, paymentDate, amount, createdAt]
      properties:
        id: { type: string, format: uuid }
        loanId: { type: string, format: uuid }
        installmentSeq: { type: integer }
        paymentDate: { type: string, format: date }
        amount: { type: number }
        transactionId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
    CreateLoanChargeRequest:
      type: object
      required: [chargeType, amount, chargeDate]
      properties:
        chargeType: { $ref: '#/components/schemas/LoanChargeType' }
        amount: { type: number }
        chargeDate: { type: string, format: date }
        notes: { type: string }
        transactionId: { type: string, format: uuid }
    LoanChargeResponse:
      type: object
      required: [id, loanId, chargeType, amount, chargeDate, createdAt]
      properties:
        id: { type: string, format: uuid }
        loanId: { type: string, format: uuid }
        chargeType: { $ref: '#/components/schemas/LoanChargeType' }
        amount: { type: number }
        chargeDate: { type: string, format: date }
        notes: { type: string }
        transactionId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
    LoansSummaryResponse:
      type: object
      required: [totalOutstanding, activeLoanCount, lentOutstanding, borrowedOutstanding, netReceivable]
      properties:
        totalOutstanding: { type: number }
        activeLoanCount: { type: integer }
        lentOutstanding: { type: number }
        borrowedOutstanding: { type: number }
        netReceivable: { type: number }
    MatchCandidateDto:
      type: object
      required: [id, date, amount]
      properties:
        id: { type: string, format: uuid }
        date: { type: string, format: date }
        amount: { type: number }
        description: { type: string }
        sourcedDescription: { type: string }
    InstallmentMatchSuggestionDto:
      type: object
      required: [installmentSeq, dueDate, expectedAmount, candidates]
      properties:
        installmentSeq: { type: integer }
        dueDate: { type: string, format: date }
        expectedAmount: { type: number }
        candidates: { type: array, items: { $ref: '#/components/schemas/MatchCandidateDto' } }
    MatchSuggestionsResponse:
      type: object
      required: [suggestions]
      properties:
        suggestions: { type: array, items: { $ref: '#/components/schemas/InstallmentMatchSuggestionDto' } }
    InstallmentDto:
      type: object
      required: [seq, dueDate, openingBalance, emi, interest, principal, closingBalance, status]
      properties:
        seq: { type: integer }
        dueDate: { type: string, format: date }
        openingBalance: { type: number }
        emi: { type: number }
        interest: { type: number }
        principal: { type: number }
        closingBalance: { type: number }
        status: { type: string }
        payment: { $ref: '#/components/schemas/LoanPaymentResponse' }
    ScheduleResultResponse:
      type: object
      required: [installments, outstandingPrincipal, totalInterestPaid, totalInterestRemaining, effectiveAprPct]
      properties:
        installments: { type: array, items: { $ref: '#/components/schemas/InstallmentDto' } }
        outstandingPrincipal: { type: number }
        totalInterestPaid: { type: number }
        totalInterestRemaining: { type: number }
        effectiveAprPct: { type: number }
...
    PageLoanResponse:
      type: object
      required: [content, number, size, totalElements, totalPages, first, last, empty]
      properties:
        content: { type: array, items: { $ref: '#/components/schemas/LoanResponse' } }
        number: { type: integer }
        size: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }
        empty: { type: boolean }
...